Add: comments to pkg/user change: make user operator private
